+++ gcs/src/main/java/com/google/cloud/hadoop/fs/gcs/GoogleHadoopFileSystem.java
@@ -91,6 +91,7 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ThreadFactory;
+import java.util.concurrent.TimeUnit;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import javax.annotation.Nonnull;
@@ -171,6 +172,8 @@ public class GoogleHadoopFileSystem extends FileSystem implements IOStatisticsSo
   static final String VERSION;
   /** Identifies this version of the {@link GoogleHadoopFileSystem} library. */
   static final String GHFS_ID;
+
+  static final String GETFILESTATUS_FILETYPE_HINT = "fs.gs.getfilestatus.filetype.hint";
 
   static {
     VERSION =
@@ -589,6 +592,7 @@ public class GoogleHadoopFileSystem extends FileSystem implements IOStatisticsSo
         });
   }
 
+  /** This is an experimental API and can change without notice. */
   public FSDataInputStream open(FileStatus status) throws IOException {
     logger.atFine().log("openWithStatus(%s)", status);
 
@@ -893,7 +897,6 @@ public class GoogleHadoopFileSystem extends FileSystem implements IOStatisticsSo
           URI gcsPath = getGcsPath(hadoopPath);
           FileInfo fileInfo = getGcsFs().getFileInfo(gcsPath);
           if (!fileInfo.exists()) {
-            GoogleCloudStorageEventBus.postOnException();
             throw new FileNotFoundException(
                 String.format(
                     "%s not found: %s", fileInfo.isDirectory() ? "Directory" : "File", hadoopPath));
@@ -901,6 +904,57 @@ public class GoogleHadoopFileSystem extends FileSystem implements IOStatisticsSo
           String userName = getUgiUserName();
           return getGoogleHadoopFileStatus(fileInfo, userName);
         });
+  }
+
+  /**
+   * Gets FileStatus with Hint. Can be used if the caller want to pass the path type (file vs
+   * directory) hint. This hint can be used to prioritize GCS API calls inorder to improve
+   * performance and reduce redundant API calls without compromising performance and API behaviour.
+   * Currently, only "file" type hint is supported.
+   *
+   * <p>This is an experimental API can can change without notice.
+   */
+  public FileStatus getFileStatusWithHint(Path hadoopPath, Configuration hint) throws IOException {
+    return trackDurationWithTracing(
+        instrumentation,
+        globalStorageStatistics,
+        GhfsStatistic.INVOCATION_GET_FILE_STATUS,
+        hadoopPath,
+        traceFactory,
+        () -> {
+          checkArgument(hadoopPath != null, "hadoopPath must not be null");
+          checkArgument(hint != null, "hint must not be null");
+
+          checkOpen();
+
+          GoogleCloudStorageFileSystemImpl.PathTypeHint pathTypeHint = getHint(hint, hadoopPath);
+          if (pathTypeHint == GoogleCloudStorageFileSystemImpl.PathTypeHint.NONE) {
+            logger.atWarning().atMostEvery(1, TimeUnit.MINUTES).log(
+                "No file type hint was provided for path %s", hadoopPath);
+          }
+
+          URI gcsPath = getGcsPath(hadoopPath);
+          FileInfo fileInfo = getGcsFs().getFileInfoWithHint(gcsPath, pathTypeHint);
+          if (!fileInfo.exists()) {
+            throw new FileNotFoundException(
+                String.format(
+                    "%s not found: %s", fileInfo.isDirectory() ? "Directory" : "File", hadoopPath));
+          }
+          String userName = getUgiUserName();
+          return getGoogleHadoopFileStatus(fileInfo, userName);
+        });
+  }
+
+  private GoogleCloudStorageFileSystemImpl.PathTypeHint getHint(Configuration hint, Path path) {
+    String hintString = hint.get(GETFILESTATUS_FILETYPE_HINT);
+    if (hintString != null && hintString.toLowerCase().equals("file")) {
+      return GoogleCloudStorageFileSystemImpl.PathTypeHint.FILE;
+    }
+
+    logger.atWarning().atMostEvery(10, TimeUnit.SECONDS).log(
+        "Unexpected hint '%s' received. Ignoring. path=%s", hintString, path);
+
+    return GoogleCloudStorageFileSystemImpl.PathTypeHint.NONE;
   }
 
   @Override
@@ -1233,6 +1287,8 @@ public class GoogleHadoopFileSystem extends FileSystem implements IOStatisticsSo
     switch (Ascii.toLowerCase(capability)) {
       case CommonPathCapabilities.FS_APPEND:
       case CommonPathCapabilities.FS_CONCAT:
+      case GcsConnectorCapabilities.OPEN_WITH_STATUS:
+      case GcsConnectorCapabilities.GET_FILE_STATUS_WITH_HINT:
         return true;
       default:
         return false;
@@ -1932,5 +1988,11 @@ public class GoogleHadoopFileSystem extends FileSystem implements IOStatisticsSo
       return String.format(
           "%s: %s", getAlgorithmName(), bytes == null ? null : BaseEncoding.base16().encode(bytes));
     }
+  }
+
+  private class GcsConnectorCapabilities {
+    public static final String OPEN_WITH_STATUS = "fs.gs.capability.open.with.status";
+    public static final String GET_FILE_STATUS_WITH_HINT =
+        "fs.gs.capability.getfilestatus.with.hint";
   }
 }