+++ gcs/CHANGES.md
@@ -1,5 +1,7 @@
 # Release Notes
 
+## Next
+
 ## 3.0.2 - 2024-08-06
 1. PR #1230 - Avoid registering subscriber class multiple times