+++ gcs/CHANGES.md
@@ -3,6 +3,8 @@
 ## Next
 1. PR #1244 - Add open() API which takes FileStatus as parameter
 
+1. PR #1253 - Add getFileStatusWithHint() API
+
 ## 3.0.2 - 2024-08-06
 1. PR #1230 - Avoid registering subscriber class multiple times